requirements.txt
FreeSimpleGUI~=5.2.0.post1
typing~=3.7.4.3
configparser~=7.2.0
pillow~=10.4.0
numpy~=1.26.4
matplotlib~=3.10.3
fitz~=0.0.1.dev2
pandas~=2.2.3
psutil~=7.0.0
pyperclip~=1.9.0
pywin32~=311
requests~=2.32.3
pydantic~=2.11.7